requirements.txt
google_api_python_client==2.127.0
protobuf==5.26.1
streamlit==1.28.1
